File structure modularized
