+++ src/events/on_close.py
@@ -1,0 +1,3 @@
+
+def on_close(ws):
+    print("- Closed connection")
