+++ src/events/on_error.py
@@ -1,0 +1,3 @@
+
+def on_error(ws, error):
+    print(error)
